LayoutTests/http/tests/security/xssAuditor/malformed-xss-protection-header-2-expected.txt

CONSOLE MESSAGE: line 1: Error parsing header X-XSS-Protection: red: expected 0 or 1 at character position 0. The default protections will be applied.
CONSOLE MESSAGE: line 4: Refused to execute a JavaScript script. Source code of script found within request.

This tests that the X-XSS-Protection header is not ignored when the first character is not 0 or 1, and that we issue an error.



--------
Frame: 'frame'
--------
If you see this message and no JavaScript alert() then the test PASSED.
